requirements.txt
appdirs==1.4.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.10
django-bootstrap3==8.2.3
django-crispy-forms==1.8.1
django-environ==0.4.5
django-filter==2.2.0
django-form-utils==1.0.3
django-froala-editor==3.1.0
django-hosts==4.0
django-jet==1.0.8
django-storages==1.8
django-tables2==1.14.2
djangorestframework==3.8.2
dropbox==8.7.1
et-xmlfile==1.0.1
flake8==3.4.1
google-api-python-client==1.7.9
google-auth==1.10.0
google-auth-httplib2==0.0.3
gunicorn==19.6.0
httplib2==0.9.2
idna==2.7
jdcal==1.4.1
Markdown==2.6.7
mccabe==0.6.1
oauth2client==4.0.0
oauthlib==2.0.2
odfpy==1.4.0
openpyxl==3.0.0
packaging==16.8
Pillow==7.1.1
pkg-resources==0.0.0
psycopg2-binary==2.8.3
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycodestyle==2.3.1
pyflakes==1.5.0
pyparsing==2.1.10
python-http-client==2.2.1
python-openid==2.2.5
pytz==2017.2
PyYAML==5.3
requests==2.20.0
requests-oauthlib==0.8.0
rsa==3.4.2
sendgrid==3.6.3
sendgrid-django==4.0.4
six==1.10.0
slackclient==1.3.0
sqlparse==0.3.0
tablib==0.12.1
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.24.3
uuid==1.30
websocket-client==0.57.0
whitenoise==3.3.1
xlrd==1.2.0
xlwt==1.3.0
